WFPREV-670 Block activity deletion and show snackbar if activity has attachments (#983)

+++ client/wfprev-war/src/main/angular/src/app/components/edit-project/activities/activities.component.ts
@@ -1,5 +1,5 @@
 import { CommonModule } from '@angular/common';
-import { ChangeDetectorRef, Component, EventEmitter, Input, OnChanges, Output, SimpleChanges, ViewChild } from '@angular/core';
+import { ChangeDetectorRef, Component, EventEmitter, Input, OnChanges, Output, QueryList, SimpleChanges, ViewChild, ViewChildren } from '@angular/core';
 import { FormBuilder, FormControl, FormGroup, FormsModule, ReactiveFormsModule, Validators } from '@angular/forms';
 import { MomentDateAdapter } from '@angular/material-moment-adapter';
 import { MatCheckboxModule } from '@angular/material/checkbox';
@@ -67,6 +67,7 @@ export class ActivitiesComponent implements OnChanges, CanComponentDeactivate {
   @Input() fiscalGuid: string = '';
   @Output() boundariesUpdated = new EventEmitter<void>();
   @ViewChild('activitiesPanel') activitiesPanel?: MatExpansionPanel;
+  @ViewChildren(ProjectFilesComponent) private attachmentFiles!: QueryList<ProjectFilesComponent>;
   messages = Messages;
   isNewActivityBeingAdded = false;
 
@@ -520,7 +521,7 @@ export class ActivitiesComponent implements OnChanges, CanComponentDeactivate {
   addActivity(): void {
     if (this.isNewActivityBeingAdded) return;
 
-    const createAndFocusNewActivity  = () => {
+    const createAndFocusNewActivity = () => {
       this.isNewActivityBeingAdded = true;
       const newActivity: ActivityModel = {};
 
@@ -542,11 +543,11 @@ export class ActivitiesComponent implements OnChanges, CanComponentDeactivate {
     if (this.activitiesPanel && !this.activitiesPanel.expanded) {
       const sub = this.activitiesPanel.opened.pipe(take(1)).subscribe(() => {
         sub.unsubscribe();
-        createAndFocusNewActivity ();
+        createAndFocusNewActivity();
       });
       this.activitiesPanel.open();
     } else {
-      createAndFocusNewActivity ();
+      createAndFocusNewActivity();
     }
   }
 
@@ -739,6 +740,18 @@ export class ActivitiesComponent implements OnChanges, CanComponentDeactivate {
           this.isNewActivityBeingAdded = false;
           return;
         }
+
+        // Block deletion if there are attachments
+        const activityAttachments = this.attachmentFiles?.toArray?.()[index];
+        if (activityAttachments?.hasAttachments) {
+          this.snackbarService.open(
+            this.messages.activityWithAttachmentDeleteFailure,
+            'OK',
+            { duration: 5000, panelClass: 'snackbar-error' }
+          );
+          return;
+        }
+
         // Delete from the service call if it's a saved fiscal activity
         this.projectService.deleteActivity(this.projectGuid, this.fiscalGuid, activityGuid)
           .subscribe({
@@ -814,4 +827,3 @@ export class ActivitiesComponent implements OnChanges, CanComponentDeactivate {
   }
 
 }
-
